Application to predict load demand
This project focus on accurately predicting electrical load demand  which I’ve developed as a web-based application with  the help of machine learning and make this forecasting process accessible, accurate, and easy to use through a website.

The application uses various machine learning techniques to generate these predictions. Users simply enter the name of a city, and the application provides a real-time prediction of the load demand for that location. It’s designed to be intuitive, so both industry professionals and the general public can benefit from the insights.

The website interface itself is powered by HTML and CSS, ensuring that it’s visually accessible and user-friendly. On the backend, I used Flask to manage data flow and load the trained machine learning models, while a JavaScript script processes the inputs and dynamically updates the forecast on the webpage.

For instance, linear regression provides reliable short-term predictions, while more advanced algorithms, such as LSTM, improve accuracy when considering additional data like weather factors—temperature, humidity, and wind speed. These weather parameters are fetched in real-time using the OpenWeatherMap API, so our predictions are both timely and context-aware.